ftn: fix the arrival time on 32-bit and on Windows
> Written by Claude (Opus 5), an LLM made by Anthropic, at Andy's
> direction.

cb1e1870 read the change time without converting it: Ctim.Sec and Ctim.Nsec
are int32 on a 32-bit build, so the vet job for the door builds this project
supports would not compile.

Windows fell through to the modification time, which is the value binkp
carries from the sender, so a freshly delivered bundle read as hours old.
Its creation time is when the file landed and os.Chtimes does not move it.
The test now probes for an arrival time rather than naming platforms, and
skips where there is none to read.

+++ internal/ftn/arrival_ctim.go
@@ -10,7 +10,9 @@ import (
 
 func arrivedAt(info os.FileInfo) time.Time {
 	if st, ok := info.Sys().(*syscall.Stat_t); ok {
-		return time.Unix(st.Ctim.Sec, st.Ctim.Nsec)
+		// The fields are int32 on a 32-bit build, so convert rather
+		// than assume the width.
+		return time.Unix(int64(st.Ctim.Sec), int64(st.Ctim.Nsec))
 	}
 	return info.ModTime()
 }
